separate fullscreen handling in viewer
